+++ src/main/java/org/truffleruby/core/binding/BindingNodes.java
@@ -391,7 +391,7 @@ public abstract class BindingNodes {
         }
     }
 
-    @CoreMethod(names = "__allocate__", constructor = true, visibility = Visibility.PRIVATE)
+    @CoreMethod(names = { "__allocate__", "__layout_allocate__" }, constructor = true, visibility = Visibility.PRIVATE)
     public abstract static class AllocateNode extends UnaryCoreMethodNode {
 
         @TruffleBoundary