+++ src/main/java/org/truffleruby/core/exception/ExceptionNodes.java
@@ -37,7 +37,7 @@ public abstract class ExceptionNodes {
 
     protected final static String CUSTOM_BACKTRACE_FIELD = "@custom_backtrace";
 
-    @CoreMethod(names = "__allocate__", constructor = true, visibility = Visibility.PRIVATE)
+    @CoreMethod(names = { "__allocate__", "__layout_allocate__" }, constructor = true, visibility = Visibility.PRIVATE)
     public abstract static class AllocateNode extends CoreMethodArrayArgumentsNode {
 
         @Child private AllocateObjectNode allocateObjectNode = AllocateObjectNode.create();