+++ src/main/java/org/truffleruby/core/fiber/FiberNodes.java
@@ -78,7 +78,7 @@ public abstract class FiberNodes {
 
     }
 
-    @CoreMethod(names = "__allocate__", constructor = true, visibility = Visibility.PRIVATE)
+    @CoreMethod(names = { "__allocate__", "__layout_allocate__" }, constructor = true, visibility = Visibility.PRIVATE)
     public abstract static class AllocateNode extends CoreMethodArrayArgumentsNode {
 
         @TruffleBoundary