+++ src/main/java/org/truffleruby/core/proc/ProcNodes.java
@@ -49,7 +49,7 @@ import com.oracle.truffle.api.source.SourceSection;
 @CoreModule(value = "Proc", isClass = true)
 public abstract class ProcNodes {
 
-    @CoreMethod(names = "__allocate__", constructor = true, visibility = Visibility.PRIVATE)
+    @CoreMethod(names = { "__allocate__", "__layout_allocate__" }, constructor = true, visibility = Visibility.PRIVATE)
     public abstract static class AllocateNode extends UnaryCoreMethodNode {
 
         @Specialization