+++ src/main/java/org/truffleruby/core/range/RangeNodes.java
@@ -519,7 +519,7 @@ public abstract class RangeNodes {
 
     }
 
-    @CoreMethod(names = "__allocate__", constructor = true, visibility = Visibility.PRIVATE)
+    @CoreMethod(names = { "__allocate__", "__layout_allocate__" }, constructor = true, visibility = Visibility.PRIVATE)
     public abstract static class AllocateNode extends UnaryCoreMethodNode {
 
         @Child private AllocateObjectNode allocateNode = AllocateObjectNode.create();