+++ src/main/java/org/truffleruby/core/tracepoint/TracePointNodes.java
@@ -38,7 +38,7 @@ import com.oracle.truffle.api.object.DynamicObject;
 @CoreModule(value = "TracePoint", isClass = true)
 public abstract class TracePointNodes {
 
-    @CoreMethod(names = "__allocate__", constructor = true, visibility = Visibility.PRIVATE)
+    @CoreMethod(names = { "__allocate__", "__layout_allocate__" }, constructor = true, visibility = Visibility.PRIVATE)
     public abstract static class AllocateNode extends UnaryCoreMethodNode {
 
         @Child private AllocateObjectNode allocateNode = AllocateObjectNode.create();